requirements.txt
torch==1.8.1
numpy==1.20.3
transformers==4.5.1
datasets==1.6.2
tiktoken==0.0.1
wandb==0.10.31
tqdm==4.60.0
Pillow==8.2.0
torchvision==0.9.1
torchscript==1.8.1
torch-model-archiver==0.4.0
